Maby fixed a bug with guild applications

+++ src/com/confusinguser/sbgods/discord/ReactionListener.java
@@ -28,6 +28,10 @@ public class ReactionListener extends ListenerAdapter {
 
     @Override
     public void onMessageReactionAdd(MessageReactionAddEvent e) {
+        for (DiscordServer server : main.getActiveServers()) {
+            if (e.getGuild().getId().contentEquals(server.getServerId())) return; //only run on active servers
+        }
+
         DiscordServer currentDiscordServer = DiscordServer.getDiscordServerFromDiscordGuild(e.getGuild());
         if ((e.getUser() != null && e.getUser().isBot()) || e.getUser() == null || currentDiscordServer == null) {
             return;
@@ -82,10 +86,12 @@ public class ReactionListener extends ListenerAdapter {
             boolean meetsSlayer = false;
             boolean meetsSkill = false;
 
-            if (player.getGuildId().equals(currentDiscordServer.getHypixelGuild().getGuildId())) {
-                e.getChannel().sendMessage(e.getUser().getAsMention() + " you are already in the guild.").complete().delete().queueAfter(30, TimeUnit.SECONDS);
-                e.getChannel().deleteMessageById(messageId).queue();
-                return;
+            if(!player.getGuildId().equals(null)) {
+                if (player.getGuildId().equals(currentDiscordServer.getHypixelGuild().getGuildId())) {
+                    e.getChannel().sendMessage(e.getUser().getAsMention() + " you are already in the guild.").complete().delete().queueAfter(30, TimeUnit.SECONDS);
+                    e.getChannel().deleteMessageById(messageId).queue();
+                    return;
+                }
             }
 
             if (slayerExp.getTotalExp() > currentDiscordServer.getHypixelGuild().getSlayerReq()) {
@@ -127,7 +133,11 @@ public class ReactionListener extends ListenerAdapter {
             e.getChannel().deleteMessageById(messageId).queue();
 
         } catch (Exception ex) {
+            main.logger.warning("Guild application failed for player " + e.getUser().getAsTag());
             main.logger.warning(ex.getMessage());
+            for(StackTraceElement msg : ex.getStackTrace()){
+                main.logger.warning("AT: " + msg.getFileName() + " " + msg.getClassName() + " " + msg.getMethodName() + " line: " + msg.getLineNumber());
+            }
             e.getChannel().sendMessage(e.getUser().getAsMention() + " there was a error somewhere, get in contact with a bot dev to help fix the error.").complete().delete().queueAfter(30, TimeUnit.SECONDS);
             e.getChannel().deleteMessageById(messageId).queue();
             return;
